refactor(xtask): rename external_bin_tests to tauri_bundle_tests
